Move `block_encoding.py` into `block_encoding/` subpackage and split into multiple files (#997)
* Move and split LCU block encoding

* Move and split Chebyshev Polynomial block encoding

* Move block_encoding.py to block_encoding/block_encoding_base.py

* Update references and move select_and_prepare.py to block_encoding/

* Add copyright notice to init file

* Fix remaining typos, imports and regenerate notebooks via autogenerate-bloqs-notebooks-v2.py
