+++ qualtran/bloqs/block_encoding/__init__.py
@@ -1,0 +1,23 @@
+#  Copyright 2024 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+r"""High level bloqs for defining bloq encodings and operations on block encodings."""
+
+from qualtran.bloqs.block_encoding.block_encoding_base import BlockEncoding
+from qualtran.bloqs.block_encoding.chebyshev_polynomial import ChebyshevPolynomial
+from qualtran.bloqs.block_encoding.lcu_block_encoding import (
+    BlackBoxPrepare,
+    BlackBoxSelect,
+    LCUBlockEncoding,
+    LCUBlockEncodingZeroState,
+)
