+++ qualtran/bloqs/block_encoding/lcu_block_encoding_test.py
@@ -11,21 +11,15 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import pytest
 
-from qualtran.bloqs.basic_gates import TGate
-from qualtran.bloqs.basic_gates.t_gate import TGate
-from qualtran.bloqs.block_encoding import (
-    _black_box_chebyshev_poly,
+from qualtran.bloqs.block_encoding.lcu_block_encoding import (
     _black_box_lcu_block,
     _black_box_lcu_zero_state_block,
     _black_box_prepare,
     _black_box_select,
-    _chebyshev_poly,
     _lcu_block,
     _lcu_zero_state_block,
 )
-from qualtran.testing import execute_notebook
 
 
 def test_lcu_block_encoding(bloq_autotester):
@@ -50,22 +44,3 @@ def test_black_box_prepare(bloq_autotester):
 
 def test_black_box_select(bloq_autotester):
     bloq_autotester(_black_box_select)
-
-
-def test_chebyshev(bloq_autotester):
-    bloq_autotester(_chebyshev_poly)
-
-
-def test_black_box_chebyshev(bloq_autotester):
-    bloq_autotester(_black_box_chebyshev_poly)
-
-
-def test_chebyshev_t_counts():
-    counts = _chebyshev_poly().call_graph()[1]
-    counts_decomp = _chebyshev_poly().decompose_bloq().call_graph()[1]
-    assert counts[TGate()] == counts_decomp[TGate()]
-
-
-@pytest.mark.notebook
-def test_notebook():
-    execute_notebook('block_encoding')
