+++ qualtran/bloqs/for_testing/qubitization_walk_test.py
@@ -20,9 +20,9 @@ import scipy
 from numpy.typing import NDArray
 
 from qualtran import Signature
+from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle
 from qualtran.bloqs.multiplexers.select_pauli_lcu import SelectPauliLCU
 from qualtran.bloqs.qubitization_walk_operator import QubitizationWalkOperator
-from qualtran.bloqs.select_and_prepare import PrepareOracle
 from qualtran.bloqs.state_preparation import PrepareUniformSuperposition
 from qualtran.symbolics import SymbolicFloat
